server/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
python-dotenv==1.0.1
